Split shell auth owner tests
